fix: Refactor frontend
