+++ frontend/package.json
@@ -19,7 +19,7 @@
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
     "@langchain/core": "^0.3.10",
-    "@langchain/langgraph-sdk": "^0.0.4",
+    "@langchain/langgraph-sdk": "^0.0.20",
     "@radix-ui/react-collapsible": "^1.1.1",
     "@radix-ui/react-dialog": "^1.1.2",
     "@radix-ui/react-dropdown-menu": "^2.1.1",
